requirements.txt
beautifulsoup4==4.13.4
Cerberus==1.3.5
cryptography==45.0.5
Flask==3.1.1
Flask_Cors==5.0.0
Flask_JWT_Extended==4.7.1
flask_limiter==3.12
numpy==2.3.1
pandas==2.3.1
Pillow==11.3.0
pycryptodome==3.22.0
pyjsparser==2.7.1
PyJWT==2.10.1
PyMySQL==1.1.1
python_bcrypt==0.3.2
python_dateutil==2.9.0.post0
pyzipper==0.3.6
Requests==2.32.4
selenium==4.34.2
selenium_stealth==1.0.6
six==1.17.0
SQLAlchemy==2.0.36
tzlocal==5.2
urllib3==2.5.0
waitress==3.0.2
Werkzeug==3.1.3
